pycon_2017/requirements.txt
Babel==2.4.0
Django==1.11.1
django-phonenumber-field==1.3.0
mysqlclient==1.3.10
phonenumberslite==8.5.0
pytz==2017.2
